requirements.txt
hikari[speedups]>=2.0.0.dev113
hikari-lightbulb>=2.3.1
hikari-miru>=2.0.1
uvloop==0.17.0

python-dotenv>=0.21.0

aiohttp==3.8.3
asyncpraw==7.6.1
plotly==5.11.0
kaleido==0.2.1
pillow==9.4.0
isodate==0.6.1
yt_dlp==2023.1.6
fast_colorthief==0.0.4
bs4==0.0.1
lxml==4.9.2
requests==2.28.1
numpy==1.24.1